feat(proto): infer ColDateTime64Auto
