+++ testdata/fuzz/FuzzDecodeBlock/0546a2c8dad97ea70bdac794ca2b7fef1765dab943a973f874d59872b44a080f
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("\x00\x00\x01\xc0\xc0\xc0\xc0\x03\x03foo\rArray(String)000000000000000000000000\x03000\x03000\x01000")
