+++ testdata/fuzz/FuzzDecodeBlock/2bddd9be45bf868704aa2612be7a95d87b408f9ed7fd629c3db475140844b732
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("\x00\x00\x01\xc0\xc0\xc0\xc0\x03\x03foo\rArray(String)000000000\x1000000000000000\x03000\x03000\x01000")
